Refactor of the app structure
