+++ src/configurations/store/configureStore.dev.js
@@ -2,8 +2,8 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import createLogger from 'redux-logger';
 
-import store from '../reducers';
-import DevTools from '../containers/DevTools.jsx';
+import store from '../../reducers';
+import DevTools from '../../containers/DevTools.jsx';
 
 const loggerMiddleware = createLogger();
 
@@ -19,8 +19,8 @@ export default function configureStore(initialState) {
     const globalStore = finalCreateStore(store, initialState);
 
     if (module.hot) {
-        module.hot.accept('../reducers', () =>
-            globalStore.replaceReducer(require('../reducers'))
+        module.hot.accept('../../reducers', () =>
+            globalStore.replaceReducer(require('../../reducers'))
         );
     }
